Feat/#6 week3 assignment

+++ src/main/java/org/sopt/service/PostService.java
@@ -2,12 +2,15 @@ package org.sopt.service;
 
 import jakarta.persistence.EntityNotFoundException;
 import org.sopt.domain.Post;
-import org.sopt.dto.request.PostRequest;
-import org.sopt.dto.request.PostUpdateRequest;
+import org.sopt.domain.User;
+import org.sopt.dto.request.post.PostCreateRequest;
+import org.sopt.dto.request.post.PostUpdateRequest;
+import org.sopt.dto.response.PostDetailResponse;
 import org.sopt.dto.response.PostResponse;
 import org.sopt.global.exception.PostErrorMessage;
 import org.sopt.mapper.PostMapper;
 import org.sopt.repository.PostRepository;
+import org.sopt.repository.UserRepository;
 import org.sopt.util.Validator;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,22 +20,28 @@ import java.util.List;
 @Service
 public class PostService {
     private final PostRepository postRepository;
+    private final UserRepository userRepository;
 
-    public PostService(final PostRepository postRepository) {
+    public PostService(final PostRepository postRepository, UserRepository userRepository) {
         this.postRepository = postRepository;
+        this.userRepository = userRepository;
     }
 
     @Transactional
-    public void createPost(final PostRequest postRequest) {
+    public void createPost(Long userId, final PostCreateRequest postCreateRequest) {
 
-        Validator.validateEmpty(postRequest.getTitle());
-        Validator.validateMaxLength(postRequest.getTitle());
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new EntityNotFoundException("사용자를 찾을 수 없습니다."));
+        Validator.validateEmpty(postCreateRequest.title());
+        Validator.validateEmpty(postCreateRequest.content());
+        Validator.validateMaxLength(postCreateRequest.title());
 
-        if (postRepository.existsByTitle(postRequest.getTitle())) {
+        if (postRepository.existsByTitle(postCreateRequest.title())) {
             throw new IllegalArgumentException(PostErrorMessage.POST_ALREADY_EXISTS_ERROR.getMessage());
         }
 
-        Post post = PostMapper.toEntity(postRequest);
+
+        Post post = PostMapper.toEntity(postCreateRequest, user);
         postRepository.save(post);
 
     }
@@ -44,20 +53,29 @@ public class PostService {
     }
 
     @Transactional(readOnly = true)
-    public PostResponse getPostById(Long id) {
+    public PostDetailResponse getPostById(Long id) {
         Post post = postRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException(PostErrorMessage.POST_NOT_FOUND_ERROR.getMessage()));
-        return PostMapper.toResponse(post);
+        return PostMapper.toDetailResponse(post);
     }
 
     @Transactional
-    public void updatePost(Long updateId, final PostUpdateRequest postUpdateRequest) {
+    public void updatePost(final Long userId, final Long updateId, final PostUpdateRequest postUpdateRequest) {
+        User author = userRepository.findById(userId)
+                .orElseThrow(() -> new EntityNotFoundException(PostErrorMessage.POST_NOT_FOUND_ERROR.getMessage()));
         Post post = postRepository.findById(updateId)
                 .orElseThrow(() -> new EntityNotFoundException(PostErrorMessage.POST_NOT_FOUND_ERROR.getMessage()));
 
         Validator.validateMaxLength(postUpdateRequest.newTitle());
         Validator.validateEmpty(postUpdateRequest.newTitle());
 
+        if (!post.getUser().equals(author)) {
+            throw new RuntimeException(PostErrorMessage.POST_UPDATE_FORBIDDEN.getMessage());
+        }
+
+        if (postRepository.existsByTitle(postUpdateRequest.newTitle())) {
+            throw new IllegalArgumentException(PostErrorMessage.POST_ALREADY_EXISTS_ERROR.getMessage());
+        }
         String newTitle = PostMapper.extractNewTitle(postUpdateRequest);
         post.updateTitle(newTitle);
     }
@@ -76,10 +94,13 @@ public class PostService {
     }
 
     @Transactional
-    public void deletePost(Long id) {
+    public void deletePost(final Long userId, Long id) {
         Post post = postRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException(PostErrorMessage.POST_NOT_FOUND_ERROR.getMessage()));
 
+        if (!post.getUser().getId().equals(userId)) {
+            throw new RuntimeException(PostErrorMessage.POST_DELETE_FORBIDDEN.getMessage());
+        }
         postRepository.delete(post);
     }
